requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.6
gunicorn==20.0.4
idna==2.10
peewee==3.14.3
Pillow==8.1.0
pyTelegramBotAPI==3.7.6
python-dateutil==2.8.1
python-dotenv==0.15.0
python-gyazo==2.1.0
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
whitenoise==5.2.0
